requirements.txt
# requirements.txt

numpy
protobuf
sentencepiece
torch
transformers
accelerate
matplotlib
scipy==1.11.3
nltk==3.8.1
openai
aiohttp
pyyaml
httpx
orjson
cohere
pyarrow
datasets
scikit-learn
chardet
streamlit
pdf2image
selenium
textract-py3
python-docx
pytesseract
PyMuPDF
beautifulsoup4
tqdm
